zephyr: arch separation for do_boot
