+++ boot/zephyr/include/do_boot.h
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2026 Espressif Systems (Shanghai) Co., Ltd.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef DO_BOOT_H_
+#define DO_BOOT_H_
+
+#include <bootutil/bootutil.h>
+
+void do_boot(const struct boot_rsp *rsp);
+
+#endif /* DO_BOOT_H_ */
