+++ docs/release-notes.d/zephyr-do_boot-arch-separation.md
@@ -1,0 +1,1 @@
+- Separate arch specific booting process to improve organization on `boot/zephyr/main.c`.
